:construction: add nut grpc proto

+++ src/i18n/locale.rs
@@ -29,11 +29,11 @@ lazy_static! {
 
 #[derive(Queryable, Serialize, Deserialize)]
 pub struct Item {
-    pub id: i64,
+    pub id: i32,
     pub lang: String,
     pub code: String,
     pub message: String,
-    pub version: i64,
+    pub version: i32,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
 }
@@ -54,11 +54,11 @@ pub trait Dao {
     fn count(&self) -> Result<i64>;
     fn all(&self, offset: i64, limit: i64) -> Result<Vec<Item>>;
     fn by_lang(&self, lang: &str) -> Result<Vec<Item>>;
-    fn by_id(&self, id: i64) -> Result<Item>;
+    fn by_id(&self, id: i32) -> Result<Item>;
     fn by_lang_and_code(&self, lang: &str, code: &str) -> Result<Item>;
-    fn delete(&self, id: i64) -> Result<()>;
+    fn delete(&self, id: i32) -> Result<()>;
     fn create(&self, lang: &str, code: &str, message: &str) -> Result<()>;
-    fn update(&self, id: i64, code: &str, message: &str) -> Result<()>;
+    fn update(&self, id: i32, code: &str, message: &str) -> Result<()>;
 }
 
 fn loop_yaml(
@@ -182,7 +182,7 @@ impl Dao for Connection {
             .load::<Item>(self)?;
         Ok(items)
     }
-    fn by_id(&self, id: i64) -> Result<Item> {
+    fn by_id(&self, id: i32) -> Result<Item> {
         let it = locales::dsl::locales
             .filter(locales::dsl::id.eq(id))
             .first::<Item>(self)?;
@@ -195,7 +195,7 @@ impl Dao for Connection {
             .first::<Item>(self)?;
         Ok(it)
     }
-    fn update(&self, id: i64, code: &str, message: &str) -> Result<()> {
+    fn update(&self, id: i32, code: &str, message: &str) -> Result<()> {
         let now = Utc::now().naive_utc();
         let it = locales::dsl::locales.filter(locales::dsl::id.eq(id));
         update(it)
@@ -219,7 +219,7 @@ impl Dao for Connection {
             .execute(self)?;
         Ok(())
     }
-    fn delete(&self, id: i64) -> Result<()> {
+    fn delete(&self, id: i32) -> Result<()> {
         delete(locales::dsl::locales.filter(locales::dsl::id.eq(id))).execute(self)?;
         Ok(())
     }
