+++ src/plugins/nut/models/log.rs
@@ -9,12 +9,12 @@ use super::Resource;
 
 #[derive(Queryable)]
 pub struct Item {
-    pub id: i64,
-    pub user_id: i64,
+    pub id: i32,
+    pub user_id: i32,
     pub level: String,
     pub ip: String,
     pub resource_type: String,
-    pub resource_id: i64,
+    pub resource_id: i32,
     pub message: String,
     pub created_at: NaiveDateTime,
 }
@@ -45,22 +45,22 @@ impl fmt::Display for Level {
 pub trait Dao {
     fn add<S: Into<String>>(
         &self,
-        user: i64,
+        user: i32,
         level: &Level,
         ip: &str,
         resource: &Resource,
         message: S,
     ) -> Result<()>;
 
-    fn all(&self, user: i64, offset: i64, limit: i64) -> Result<Vec<Item>>;
+    fn all(&self, user: i32, offset: i64, limit: i64) -> Result<Vec<Item>>;
     fn by_resource(&self, resource: &Resource) -> Result<Vec<Item>>;
-    fn count(&self, user: i64) -> Result<i64>;
+    fn count(&self, user: i32) -> Result<i64>;
 }
 
 impl Dao for Connection {
     fn add<S: Into<String>>(
         &self,
-        user: i64,
+        user: i32,
         level: &Level,
         ip: &str,
         resource: &Resource,
@@ -79,7 +79,7 @@ impl Dao for Connection {
         Ok(())
     }
 
-    fn all(&self, user: i64, offset: i64, limit: i64) -> Result<Vec<Item>> {
+    fn all(&self, user: i32, offset: i64, limit: i64) -> Result<Vec<Item>> {
         let items = logs::dsl::logs
             .filter(logs::dsl::user_id.eq(user))
             .order(logs::dsl::created_at.desc())
@@ -96,7 +96,7 @@ impl Dao for Connection {
             .load::<Item>(self)?;
         Ok(items)
     }
-    fn count(&self, user: i64) -> Result<i64> {
+    fn count(&self, user: i32) -> Result<i64> {
         let it = logs::dsl::logs
             .filter(logs::dsl::user_id.eq(user))
             .count()
